vimrc: figured out how to set b:did_ftplugin properly
I did need to set `b:did_ftplugin` in an autocmd. Without doing so I
was only setting it for the first opened buffer, and subsequent buffers
would have filetype plugins enabled. I'm surprised that I didn't realise
it until today. Figured out that I should be setting the autocmd on
BufReadPre in order to get the variable set properly for all buffers
that I open.