Add custom plugin better_autoread
Make a plugin out of a Stack Overflow post for a better autoread
augroup.

The `autoread` setting in vim will automatically reload a file if it's
been changed outside of vim (if you've made unsaved changes vim will
prompt before loading the file). By default, however, vim doesn't
automatically detect when a file has changed. It only happens when you
do certain things. This sets up an autocmd to force a check a bit more
often by tying into the CursorHold events. A bit hackish it seems but it
works like a charm.

This has been sorely needed especially since yesterday when I was
switching git branches in a project several times and editing various
files only to be alerted upon saving that the file on disk has changed.
Previously I would say it was a minor annoyance to have to type :e! in
those situations. Yesterday it was a freaking pain in the glaiven since
it happened unexpectedly so many times and after I was ready to
commit/save edits.