Add ag.vim plugin
Was doing a search for a class in HTML files and realised that I want
the files open in vim and in order to do so I would either have to
monkey with some shell piping or Control-Z/fg back and forth from ag
output to vim whilst remembering which filenames I wanted to open and
open them manually.

Craziness.

Finally installing Ag.vim for these instances when I want to instantly
open vim buffers for the files in my search results.