ftplugin/gitcommit.vim: Don't start wrapped lines with '*'
When making bulleted lists in git commits, I normally use '*' as the
bullet character.

When my text wrapped from the textwidth=72 setting for gitcommit-type
files, the line following the start of a bullet would begin with a
bullet.

Obviously this is the wrong behaviour since I'm continuing a list item,
not of starting a new one.

This frustrated me to no end whenever I wrote a list in a git commit,
constantly having to `0r<space>` to fix the extra asterisks.

I finally decided to solve this problem once and for all. Read up a
little on:
* :h formatoptions
* :h fo-table
* :h format-comments

The last of which gave me what I needed to fix the problem. Now
asterisks in git commits behave as list bullets.