vimrc: set sessionoptions-=options
This setting tells vim not to save options & mappings in persisted
sessions. This allows open buffers, windows, and tabs to be saved but
configuration options and plugins to be different when reopening a
session. Useful if I've made modifications to my vimrc after creating a
session.

Found out about this from https://github.com/xolox/vim-session. Really
cool that this exists. It's been something that annoyed me for a while
but I never bothered to find a solution.