virmc: Set backupcopy=yes
Add setting to preserve file attributes, such as creation date. I went
to check the creation date of a file I know was created at least a year
ago, but its creation date was today and the same as its modified date.

I had recently edited the file in Vim, and that ended up being the
problem.

This setting ensures that files are reused so their attributes are
preserved.

What an unfortunate loss of history since last April.