vimrc: Add <leader>or mapping
Add a mapping to open the current file in `ranger`, the console file
explorer. I already have a similar command to open the current file in
Finder, but these days I find ranger a bit more pleasant to use,
particularly if I'm already in the console. It's definitely more
keyboard-friendly.

This came about because I was trying to copy a file and the built-in Vim
methods for doing so that I know of were more complicated than I wanted.

1. Use :saveas [filename]

   This method does the right thing in copying but the copy goes to
   Vim's :pwd. I wanted it to go to the same directory the file was in.

2. Use Netrw

   The mf mt mc commands were a little confuddling to me. It makes sense
   when copying a file to a directory, but how do I duplicate a file in
   the same directory?

3. Use :shell

   Opens a shell, but in Vim's :pwd, not in the current file's
   directory. Maybe I should make a mapping for this.