Remove 'EasyGrep' plugin
No longer using this one. I've been using ag exclusively, mostly
directly from the console but sometimes using the plugin. For find and
replace I've gotten more adept at using the arglist, so I haven't used
EasyGrep in a long time. It can also be kind of slow, mostly because it
doesn't use ag. Not sure if there's a way to integrate the two.