vimrc: Disable auditory.vim
Wish there was an easy way to toggle the plugin on and off with a
command in the plugin (e.g. `:AuditoryToggle`). Would be nice to enable
it when I wanted to and have it disabled at other times. I like it but
there are times when I don't want my computer to be playing music when
I'm typing or editing text. Disable it by default.