vimrc: Add mapping to `git blame` current buffer
Inspired by a mapping I happened upon in Ben Orenstein's vimrc, this
will open a new vertical split with the `git blame` of the current
buffer. Still needs to be refined, but this should be useful without
having to install Fugitive (as I'm not convinced I need most of what the
plugin provides).

Noticed recently that I've been `git blame`ing a file I have open in
Vim, and need to type out the file path into a terminal in order to get
to it. Would be much quicker to just open this in Vim.