Add 'vim-grappele' plugin
This will finally give me a way to repeat `G` commands. Should be useful
since I often find myself entering a `G` command only to realise that I
was in the wrong window and have to then type the whole thing again.
With this plugin all I have to do is type a quick mapping to make the
repeat happen.