vimrc: Add a mapping to open the current file in Chrome
This is to make it easier to preview HTML files in a browser. Otherwise
I would either open the file in Finder with `<leader>of` and drag it to
a browser in my Dock or open a new terminal, `cd` into the right
directory, start a Python `SimpleHTTPServer`, and finally navigate to
`localhost:8000/___` in a browser. This makes doing the first option
quick and easy.

Since we're opening the file directly, this will use the `file://`
protocol, but that's enough to be able to take a look at small pages.