vimrc: Output `scrollbind` value with <leader>sb
Since our `<leader>sb` mapping is a toggle command for `scrollbind`,
it can be hard to know whether it's on or off after invoking it if you
haven't been paying attention. Output `setl scrollbind?` after
performing the toggle so that the current `scrollbind` value displays in
the command line area to make it clear what happened.