Add command to post mentor logs to The Firehose Project
It finally became tedious to have to manually execute my custom script
that posts a mentor log to The Firehose Project in a new shell. Create a
new command that will run the script directly within Vim so I don't have
to leave the editor.

Actual file paths are stored in `projects/file-paths.vim`.

Need to set the `BUNDLE_GEMFILE` so that the Ruby gems required by the
script get loaded.

The absolute path of the current file gets passed to the script for
posting.