Disable Slimv plugin
I've had the plugin disabled but uncommitted for many months now. I
think it's finally time to make the change official.

Slimv is an excellent plugin, and makes it possible to have a good Common
Lisp development environment in Vim. I'll still be using it for that
reason.

The trouble is that Slimv isn't a good plugin citizen: It overrides user
defaults. I noticed that at some point many months ago, code I was
writing was getting auto-indented. The "smart" kind of autoindent, not
`set autoindent`. I _hate_ "smart" auto-indenting. Figured out that
Slimv was the cause and immediately disabled the plugin.

We may want to create a Bash alias that we can use exclusively for CL
development that loads Slimv into the runtimepath instead of having it
auto-loaded by Pathogen.