git-blamer: Open blame windows `leftabove`
Based on my current Vim settings, splits open to the right (I have
`splitright` turned on).

This causes the blame window to open to the right of the file in
question, which is weird, since normally that metadata appears to the
left of the file.

Open the Blamer split on the left by default to make the formatting more
consistent with the normal `git blame` output.