Add ftplugin/todo.vim
Special overrides for *.todo files. Here we unlet `b:did_ftplugin` to
allow the plugin's ftplugin to execute. Normally, we override
`b:did_ftplugin` to disable filetype functionality. In this case,
though, we do want the ftplugin code that comes with vim-twodo to be
loaded and useable.

We also set a custom colour for regular todos that fits with our default
twilight256 colour scheme.