Add ftplugin/netrw.vim
Override the `g:netrw_bufsettings` variable to make Netrw always open
with line numbers. This should make it easier to navigate to a specific
file in the tree.

Copied the settings from the original value and changed `nonu` to `nu`.

Following this Stack Overflow answer from Benoit to get this working:
http://stackoverflow.com/questions/8730702/how-do-i-configure-vimrc-so-that-line-numbers-display-in-netrw-in-vim/8731175#8731175