Create new filetype plugin 'htmlerb'
I want the HTML matchit command to be available in *.html.erb files.
Normally, these files get loaded as `filetype=eruby`.

The trouble with that is that not all eruby is HTML, so I can't load the
matcher in the 'eruby' filetype because it might not make sense.

Create a new special filetype for *.html.erb files that uses the normal
eruby syntax, but includes the HTML matchit command.