Add 'vim-argwrap' plugin
Was getting annoyed at having to manually split function arguments from
a single line to multiple lines. Planned on writing a little custom
plugin to do this, but this one seems more well thought out and
featureful than what I was thinking of. It also seems lighter and not as
presumptive as another alternative I found:
https://github.com/AndrewRadev/splitjoin.vim