vimrc: Make <C-w>q command a no-op
When hitting <C-w> to initiate a window command, I kept accidentally
hitting <C-w>q and quitting the window I was in. This was extremely
aggravating, because (1) it would be surprising and confusing for a
couple seconds, and (2) I'd have to stop what I was doing and get my
workspace back to what it was before.

I was accidentally hitting it because the 'w' and 'q' keys are so close
together.

The only difference I can see between <C-w>q and <C-w>c is that the 'q'
version will quit Vim if the current window is the last one. I don't see
myself needing that functionality with 'ZQ' readily available. Since
closing a window is a destructive operation, I prefer it being farther
away on the keyboard (even though 'c' is right next to 'v' and that
confusion has messed me up sometimes).

All that said, get rid of this command that has a more useful
equivalent, that I don't see myself using, and that has messed me up at
least three times already.