vimrc: Add ZA mapping to update all buffers and quit
Tested this out for a while without committing it and it turns out I
actually started using this mapping frequently enough, so I'm saving it.

The real impetus for this is that I recently (in the last couple months)
set:

    $ git config commit.verbose 1

to get a diff of changes to be committed in my Vim commit buffer. This
is really handy. Often I end up opening Vim windows to see the rest of
the diff while still having my commit message open and visible. Thus my
normal workflow of writing a commit message and committing with ZZ no
longer works in these cases where there are multiple windows open. To
more quickly save the commit message and quit all windows, this command
comes in really handy.