vimrc: Log statistics of RSpec command usage
When any of my RSpec commands are used, log a message to a file that
says which one was just used.

These statistics will be used for an experiment, gathering frequency
data over time. That data will then be used to determine how often the
lowercase variants are used, and how often the uppercase variants are
used. If uppercase occurs more often than lowercase, the commands should
be swapped, because lowercase takes less time to type.