ftplugin/ruby.vim: Add `RubyNewHashSyntax` command
I've had to use a similar pattern a couple of times in the past to
convert files or sections of files from the old Ruby hash syntax
(`:key => 'value'`) to the new syntax (`key: 'value'`).

Give the pattern and substitution a command to make it readily available
the next time I need to use it.