ftplugin/ruby.vim: Add mapping to reset the Rails test database
Occasionally I need to drop and reload the Rails test database. Recently
I've been working on something that requires a "truncation"-cleaned
test, and have had to reset my database regularly while writing the test
to eliminate inconsistent state. This command gives me the ability to
reset with a single key press. I've been testing it out for a few days
in a Vim session and have found it very helpful.