projects/af83.vim: Add abbreviations for Ouibus
I'm not supposed to write this word, so change it to the code word
automatically if I happen to write it accidentally in a commit message.

Interestingly, the bars (`|`) need to go either right next to the
backslashes (`\`) or right after the right-hand side, as in:

    inoreabbrev <buffer> Ouibus the socle commun|

Otherwise, if a space is before the `|`, that space becomes part of the
abbreviation's right-hand side.