git-blamer: Align cursor lines of buffer and GitBlamer window
Take the mechanism that Fugitive.vim uses:
https://github.com/tpope/vim-fugitive/blob/f3ccb0c12ee4985b8808f83059830a24cc92821c/plugin/fugitive.vim#L2048-L2064

This positions the cursor lines on the same horizontal line so that
blame information can be read on the same line as the code it
corresponds to.

Not sure, but it doesn't seem to work exactly right on one of my
machines, but did appear to work correctly on another. Could just be the
fault of 'wrap' though.