Add 'github-url' plugin
A rough working implementation of a plugin that generates a GitHub URL
for the current file.

It gets the base URL from the current Git repo's `origin` remote and
uses the current commit SHA. Lines can also be highlighted by passing a
range to the command.

This makes it easier to share a bit of code that I see in the editor
with other people on my team in chat or the issue tracker.