vimrc: Add basic 'vim-go' settings
'vim-go' displays this error message when running older versions of Vim:

    vim-go requires Vim 7.4.2009 or Neovim 0.3.1, but you're using an older version.
    Please update your Vim for the best vim-go experience.
    If you really want to continue you can set this to make the error go away:
        let g:go_version_warning = 0
    Note that some features may error out or behave incorrectly.
    Please do not report bugs unless you're using Vim 7.4.2009 or newer or Neovim 0.3.1.O

If I'm running a version of Vim that would cause the error message to
display, I'm probably not editing Go code, so it doesn't matter. In this
case, the error message just interferes with normal operations.

Use `goimports` instead of `go fmt` for on-save formatting.