plugin/better_autoread.vim: Disable automatic `checktime` on Vim 8
I kept getting E211 errors when switching Git branches with
added/missing files or E321 errors. Normally, I'd get the error once,
dismiss it, and be done with it, but on a new machine with a recent Vim
install, the errors kept coming back up every few seconds,
insupportable.

Unless I can think of something better, take a heavy-handed approach and
disable this on new Vims.