diff_corrections: Add `cx` mapping to close tab
Add a mapping to quickly close a tab. Makes it easier to close files
that I've finished reviewing.

Thanks to user9433424
(https://vi.stackexchange.com/users/6960/user9433424) on the Vi Stack
Exchange for the very detailed explanation of how to restore a mapping:
https://vi.stackexchange.com/questions/7734/how-to-save-and-restore-a-mapping/7735#7735

Using user9433424's mapping restore code, and ignoring the case where a
mapping is defined both globally and local to a buffer.