Add 'jira-text-formatting-notation-help' plugin
Plugin that opens a window containing the text formatting guide for
Jira.

The guide is taken from:

    $ w3m -dump 'https://jira.atlassian.com/secure/WikiRendererHelpAction.jspa?section=all'

Since I'm using the go-jira CLI, I need the reference for any special
formatting I want to use. I figured it would help to have that
information right next to my Vim buffer instead of having to switch to a
browser and open the web page.