ftplugin/go.vim: Add mapping to run package tests with terminal output
The Vim-Go mappings run the tests in a job and add results to the
Quickfix list. Add a new mapping that doesn't use Quickfix, and instead
just runs the test command for the package that contains the current
file, outputting the results to standard output.

Gives me an additional alternative to view test results.