ftplugin/go.vim: Create test file on alternate if it doesn't exist
When switching to a file's alternate (its test file), create the test if
the file doesn't exist. By default, a warning is issued if the file
doesn't exist. I wanted an easier way to create the test file.

Previously I'd use:

    :e %:h<Tab><C-f>bbi_test<CR>

This uses the function that the `Plug` mappings call under the hood, and
turns on the `-bang` argument to have it create the test file if it
doesn't exist.