vimrc: Set a fixed tall quickfix window height for vim-go
By default, vim-go sets the quickfix window's height to be the number of
quickfix items. This is problematic when, for example, there's a single
error item, and the quickfix window opens below a narrow window (for
example, in one of three vertical splits). In that case, the error gets
line wrapped, and quickfix window is only one line tall, meaning I can't
read the full error.

Set the window height to Vim's default of 10 lines. Vim-go will collapse
the height to 10 lines if the variable is set to a number greater than
10, and 5 lines was too short in practise.