vimrc: Replace `t` cabbrev with `T` command alias
Alias `T` to `TComment` instead of using a cabbrev.

The abbreviation would mess up commands with "t" in them, like turning
`tjump` into `TCommentjump`.

Replace the abbreviation with a new `T` command that acts as an alias
for `TComment`. I didn't include argument completion because I don't use
that. Otherwise, replicate the `TComment` command's definition and pass
all parameters to it. This gives us a short alias for commenting, and
`T` is the first completion with `:t<Tab>`.