projects/aodocs.vim: Make `s:TodoCopyLast()` work for second entry
When only a single TODO day entry exists in a file, this function would
fail with a "pattern not found" error because the copy pattern expects
at least two entries in a file. Alter the function so that it also works
when only one entry is present in the file.