vimrc: Add `m,` mapping to save session
I've been testing this binding as a candidate for a potential
replacement of my existing `<leader>mk` mapping. It's easier on the
hands than the existing one.

Since `,` doesn't seem to be a valid mark location, and it's right next
to the "m" key on a Qwerty keyboard, this seemed like an ideal
combination of keys for a relatively frequent action.