Add 'insert-layout' plugin idea
This uses `xkbswitch` to change keyboard layouts on insert enter and
leave. For now it's only a proof of concept.

The plan is to make an `:InsertLayout` command that takes an argument
defining the keyboard layout to use in insert mode. When returning to
normal mode, it should switch back to the previous layout.

This uses https://github.com/myshov/xkbswitch-macosx/.