Add 'prr' plugin @ 9b44c3f2d4b79013a00aae82cb59fe63306967ca
Provides syntax highlighting for Prr code review files.

This code was copied from the 'prr' repository
(https://github.com/danobi/prr/) at commit
9b44c3f2d4b79013a00aae82cb59fe63306967ca
(https://github.com/danobi/prr/commit/9b44c3f2d4b79013a00aae82cb59fe63306967ca).

The Vim plugin lives in a subdirectory of the 'prr' repository, which
made it tricky to include as a submodule.

I added GPLv2 license headers based on the license of the Prr project,
however the Vim code seems to indicate it was based on Vim-licensed
code, so I'm not clear on what the actual license is supposed to be.
I'm assuming it's GPL.