after/compiler/typescript.vim: Compile all files
The 'typescript.vim' plugin has the 'makeprg' compile only the current
file, not all files. Copy its 'makeprg' command and modify it to compile
all files in the project instead. This way I get compiler errors in
quickfix for the whole project.

This way makes more sense to me and works for code reviews.