vimrc: Add `[b` and `]b` buffer cycling mappings
I've had these on `<leader>k` and `<leader>j` since the early days, but
those mappings no longer feel logical to me. It seems that `[b` and `]b`
are unbound, so I decided to bind these to buffer switching, for the
rare times when I need it.

Eventually I'll probably remove one or the other set depending on what I
end up sticking with.