projects/aodocs.vim: Add ESLint-compatible JS debug mappings
AODocs projects use ESLint standard settings for the most part which
disallows `debugger` statements. Since it can be difficult to change
this setting at a project level, update my debugger mappings to force
enable the `debugger` statement that's inserted.

Note: This is slow. We should update it to use Vim script.