ftplugin/dart.vim: Rough draft debugger mapping
It seems you can't `import` within a block in Dart, so had to finagle a
mapping that puts the import statement at the start of the file. Not
ideal as there should only be one undo action to remove the debugger
insert, but it works for now until I can write something cleaner.