diff-corrections: Fix highlight restore error
While debugging an issue with 'colorscheme' in Git commits on one of my
machines using these 'verbose' lines in my 'vimrc':

    set verbose=9
    set verbosefile=/tmp/vimverbose.txt

I discovered the following error coming from the highlight restore
autocmd in this plugin:

    Error detected while processing BufEnter Autocommands for
    "*"..function diff_corrections#Run: line   12: E416: Missing equal
    sign: 2: ^Isilent highlight Comment E492: Not an editor command:
    Comment term=bold ctermfg=7 guifg=#80a0ff^@^ILast set from
    ~/.vim/colors/twilight256.vim line 325^@line 3: ^Iredir END

It runs on 'BufEnter', so basically all the time. The issue is that
there's a second line that wasn't expected.

First change the `split` call to keep only the first line, which is the
one we need to execute to restore the highlight. Now we no longer need
to remove the first character from the first line for some reason.

Also add a `s:should_restore` variable so that we don't try to
incorrectly restore the highlight when it hasn't been modified.